extract dataflow_builder.rs
